+++ tests/Models/Comment.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Marifyahya\EloquentFilter\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Comment extends Model
+{
+    protected $table = 'comments';
+    protected $guarded = [];
+    public $timestamps = false;
+}